First iteration of script to store data locally.
"covid data getter" is meant to be run on a cron job that perpetually checks against a local CSV for blank values and fills them, primarily for the purpose of accessing the data more easily later. Eventually, there will be a second script that checks the local database and compiles and emails the report to subscribers, and eventually eventually, a way to generate visual charts locally without having to rely on google docs.